[Refactor] #141 App.jsx 인증/룰렛 훅 분리, 공통 모듈 추출, api.js 401 처리
- utils/useAuth.js 신설: 인증 상태머신 + 카카오/애플/자동로그인 3중복 후처리를 applySession 주입으로 통합, 401 세션 만료 시 자동 로그아웃
- utils/useMissionRouletteTrigger.js 신설: 1초 인터벌 룰렛 트리거 분리 (같은 분·hasMission 정책 유지)
- utils/api.js: 토큰 로깅 __DEV__ 가드, setOnUnauthorized 콜백 추가
- utils/date.js pad 통합(10곳), components/AvatarImage.jsx(2곳), components/WheelPicker.jsx(2곳) 중복 제거
