+++ screens/ProofDetailScreen.jsx
@@ -10,6 +10,7 @@ import {
   RoomTopBar, MemberAvatar, ReactionBar, ProofStatusBadge, OutlineButton,
 } from '../components/RoomBits';
 import ReportReasonModal from '../components/ReportReasonModal';
+import { pad } from '../utils/date';
 
 function resolvePhoto(url) {
   if (!url) return null;
@@ -22,8 +23,7 @@ function timeLabel(createdAt) {
     ? new Date(createdAt[0], (createdAt[1] || 1) - 1, createdAt[2] || 1, createdAt[3] || 0, createdAt[4] || 0)
     : new Date(createdAt);
   if (isNaN(d.getTime())) return '';
-  const p = (n) => String(n).padStart(2, '0');
-  return `${p(d.getHours())}:${p(d.getMinutes())}`;
+  return `${pad(d.getHours())}:${pad(d.getMinutes())}`;
 }
 
 export default function ProofDetailScreen({ roomId, proofId, missionTitle, isGroup = true, onBack, onChanged }) {
