+++ utils/date.js
@@ -1,0 +1,4 @@
+/* 공통 날짜/시간 유틸 */
+
+/** 2자리 zero-pad — 8 → '08' */
+export const pad = (n) => String(n).padStart(2, '0');
